made message  form into own component
