+++ app/assets/stylesheets/components/pool/chat_box.scss
@@ -23,7 +23,7 @@
     // overflow-y: auto;
 
     .chat-box-item {
-      margin-bottom: 10px;
+      margin-bottom: 20px;
       display: flex;
       flex-direction: row;
       align-items: flex-start;
