+++ frontend/modules/pool/subcomponents/index.js
@@ -2,3 +2,4 @@ export * from './chat_box_item';
 export * from './chat_box';
 export * from './mod_board_item';
 export * from './mod_board';
+export * from './message_form';
